+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 * **New Features:**
 * **Notable Changes:**
+    * hyperized Boolean AND, OR operators
 * **Notable Fixes:**
     * preserve contents of variadic inputs when editing or translating a custom blocks
 * **Documentation Updates:**
@@ -12,6 +13,7 @@
 ### 2022-02-09
 * new dev version
 * blocks, byob: preserve contents of variadic inputs when editing or translating a custom blocks
+* threads: hyperized Boolean AND, OR operators
 
 ## 7.1.3:
 * **New Features:**